refactoring: split bin.lua into seveal files
